% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apf.wsum.R
\name{apf.wsum}
\alias{apf.wsum}
\title{Weighted Sum of Multiple Accumulative Persistence Functions}
\usage{
apf.wsum(alist, weight)
}
\description{
Weighted Sum of Multiple Accumulative Persistence Functions
}
