% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2apf.R
\name{d2apf}
\alias{d2apf}
\title{Convert Persistence Diagram to Accumulated Persistence Function}
\usage{
d2apf(diagram, dimension = 1, tseq)
}
\description{
Convert Persistence Diagram to Accumulated Persistence Function
}
