% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2landscape.R
\name{d2landscape}
\alias{d2landscape}
\title{Convert Persistence Diagram to Persistence Landscape}
\usage{
d2landscape(
  diagram,
  dimension = 1,
  k = 0,
  tseq,
  inf.replace = FALSE,
  inf.repval = 2 * max(diagram[!is.infinite(diagram[, 3]), 3])
)
}
\description{
Convert Persistence Diagram to Persistence Landscape
}
\references{
\insertRef{bubenik_persistence_2018}{TDAkit}
}
