% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2silhouette.R
\name{d2silhouette}
\alias{d2silhouette}
\title{Convert Persistence Diagram to Persistence Silhouette}
\usage{
d2silhouette(diagram, dimension = 1, p = 2, tseq)
}
\description{
Convert Persistence Diagram to Persistence Silhouette
}
