% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram_diagAS.R
\name{diagAS}
\alias{diagAS}
\title{Compute Alpha Shape Persistence Diagram in 3D}
\usage{
diagAS(data, maxdim = 1)
}
\arguments{
\item{data}{an \eqn{(n\times 3)} data matrix.}

\item{maxdim}{maximum dimension of the computed homological features (default: 1).}

\item{threshold}{maximum value of the filtration (default: \code{Inf}).}
}
\value{
a data frame with following columns\describe{
\item{Dimension}{dimension corresponding to a feature.}
\item{Birth}{birth of a feature.}
\item{Death}{death of a feature.}
}
}
\description{
\code{diagAS} computes the persistent diagram of the Alpha Shape filtration 
constructed on a point cloud in \eqn{\mathbf{R}^3}. This function is a 
wrapper to \pkg{TDA}'s implementation. It is noted that AS filtration is \emph{only} 
available in 3-dimensional point cloud.
}
\examples{
# ---------------------------------------------------------------------------
#                 Compare VR and AS Persistence Diagrams
# ---------------------------------------------------------------------------
# Use 'iris' data for the first 3 columns
XX = as.matrix(iris[,1:3])

# Compute VR and AS Diagram
run.vr = diagRips(XX, maxdim=1)
run.as = diagAS(XX, maxdim=1)

col1 = as.factor(run.vr$Dimension)
col2 = as.factor(run.as$Dimension)

# Visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,2), pty="s")
plot(run.vr$Birth, run.vr$Death, pch=19, col=col1, main="VR diagram")
plot(run.as$Birth, run.as$Death, pch=19, col=col2, main="AS diagram")
par(opar)

}
\seealso{
\code{\link[TDA]{alphaShapeDiag}}
}
\concept{diagram}
