% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.2circles.R
\name{gen.2circles}
\alias{gen.2circles}
\title{Generate Two Intersecting Filled Circles}
\usage{
gen.2circles(n = 496, sd = 0)
}
\description{
Generate Two Intersecting Filled Circles
}
\examples{
## see two sets of intertwined circles
nn = 200
x1 = gen.2circles(n=nn)
x2 = gen.2circles(n=nn, sd=0.05)
x3 = gen.2circles(n=nn, sd=0.1)
cols = as.factor(c(rep(1,nn/2),rep(2,nn/2)))

## visualize
opar = par(mfrow=c(1,3), pty="s")
plot(x1[,1],x1[,2],pch=18,main="sd=0.00",col=cols)
plot(x2[,1],x2[,2],pch=18,main="sd=0.05",col=cols)
plot(x3[,1],x3[,2],pch=18,main="sd=0.10",col=cols)
on.exit(par(opar))

}
