% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.eqdist.R
\name{landscape.eqdist}
\alias{landscape.eqdist}
\title{Energy-based Testing Equality of Distributions for Persistence Landscapes}
\usage{
landscape.eqdist(
  dlist,
  label,
  method = c("original", "discoB", "discoF"),
  mc.iter = 496
)
}
\description{
Energy-based Testing Equality of Distributions for Persistence Landscapes
}
