% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.norm.R
\name{landscape.norm}
\alias{landscape.norm}
\title{\eqn{L_p} norm of a single Persistence Landscape}
\usage{
landscape.norm(landscape, p = 2)
}
\description{
\eqn{L_p} norm of a single Persistence Landscape
}
\references{
\insertRef{bubenik_persistence_2018}{TDAkit}
}
