% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.norm2.R
\name{landscape.norm2}
\alias{landscape.norm2}
\title{Pairwise \eqn{L_p} norm of multiple Persistence Landscapes}
\usage{
landscape.norm2(dlist, p = 2, as.dist = FALSE)
}
\description{
Pairwise \eqn{L_p} norm of multiple Persistence Landscapes
}
