% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.normpair.R
\name{landscape.normpair}
\alias{landscape.normpair}
\title{Pairwise \eqn{L_p} norm of multiple Persistence Landscapes}
\usage{
landscape.normpair(dlist, p = 2, as.dist = FALSE)
}
\description{
Pairwise \eqn{L_p} norm of multiple Persistence Landscapes
}
