% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.wsum.R
\name{landscape.wsum}
\alias{landscape.wsum}
\title{Weighted Sum of Multiple Persistence Landscapes}
\usage{
landscape.wsum(dlist, weight)
}
\description{
Weighted Sum of Multiple Persistence Landscapes
}
\references{
\insertRef{bubenik_persistence_2018}{TDAkit}
}
