% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notyet_napf.mean.R
\name{napf.mean}
\alias{napf.mean}
\title{Mean of Multiple Multiple Normalized Accumulative Persistence Functions}
\usage{
napf.mean(flist)
}
\description{
Mean of Multiple Multiple Normalized Accumulative Persistence Functions
}
