% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notyet_napf.wsum.R
\name{napf.wsum}
\alias{napf.wsum}
\title{Weighted Sum of Multiple Normalized Accumulative Persistence Functions}
\usage{
napf.wsum(flist, weight)
}
\description{
Weighted Sum of Multiple Normalized Accumulative Persistence Functions
}
