% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_dist.R
\name{pldist}
\alias{pldist}
\title{Pairwise \eqn{L_p} Distance of Multiple Persistence Landscapes}
\usage{
pldist(dlist, p = 2, as.dist = FALSE)
}
\arguments{
\item{dlist}{a length-\eqn{N} list of \code{'landscape'} objects from \code{\link{diag2landscape}}.}

\item{p}{an exponent in \eqn{[1,\infty)} (default: 2).}

\item{as.dist}{logical; if TRUE, it returns dist object, else it returns a symmetric matrix.}
}
\value{
a S3 \code{dist} object or \eqn{(N\times N)} symmetric matrix of pairwise distances according to \code{as.dist} parameter.
}
\description{
Given multiple landscape functions \eqn{\Lambda_1 (t), \Lambda_2 (t), \ldots, \Lambda_N (t)}, 
compute \eqn{L_p} distance in a pairwise sense where 
\deqn{\|\Lambda_i (t) - \Lambda_j (t) \|_p = \left( \sum_{k=1}^K \| \Lambda_{k,i} (t) - \Lambda_{k,j} (t) \|_p^p \right)^{1/p} } 
for top \eqn{K} landscape functions.
}
\examples{
# ---------------------------------------------------------------------------
#         Compute L_2 Distance for 3 Types of Landscapes
#
# We will compare dim=0,1 with top-5 landscape functions with 
# - Class 1 : 'iris' dataset with noise
# - Class 2 : samples from 'gen2holes()'
# - Class 3 : samples from 'gen2circles()'
# ---------------------------------------------------------------------------
## Generate Data and Diagram from VR Filtration
ndata     = 10
list_rips = list()
for (i in 1:ndata){
  dat1 = as.matrix(iris[,1:4]) + matrix(rnorm(150*4), ncol=4)
  dat2 = gen2holes(n=100, sd=1)$data
  dat3 = gen2circles(n=100, sd=1)$data
  
  list_rips[[i]] = diagRips(dat1, maxdim=1)
  list_rips[[i+ndata]] = diagRips(dat2, maxdim=1)
  list_rips[[i+(2*ndata)]] = diagRips(dat3, maxdim=1)
}

## Compute Persistence Landscapes from Each Diagram with k=5 Functions
#  We try to get distance in dimensions 0 and 1.
list_land0 = list()
list_land1 = list()
for (i in 1:(3*ndata)){
  list_land0[[i]] = diag2landscape(list_rips[[i]], dimension=0, k=5)
  list_land1[[i]] = diag2landscape(list_rips[[i]], dimension=1, k=5)
}

## Compute L2 Distance Matrix
dmat0 = pldist(list_land0, p=2)
dmat1 = pldist(list_land1, p=2)

## Visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,2), pty="s")
image(dmat0[,(3*(ndata)):1], axes=FALSE, main="distance : dim=0")
image(dmat1[,(3*(ndata)):1], axes=FALSE, main="distance : dim=1")
par(opar)

}
\concept{landscape}
