% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.median.R
\name{silhouette.median}
\alias{silhouette.median}
\title{Geometric Median of Multiple Persistence Silhouettes}
\usage{
silhouette.median(
  slist,
  weight = rep(1/length(slist), length(slist)),
  maxiter = 496,
  abstol = 1e-06,
  print.progress = TRUE
)
}
\description{
Geometric Median of Multiple Persistence Silhouettes
}
