% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.norm.R
\name{silhouette.norm}
\alias{silhouette.norm}
\title{\eqn{L_p} norm of a single Persistence Silhouette}
\usage{
silhouette.norm(silhouette, p = 2)
}
\description{
\eqn{L_p} norm of a single Persistence Silhouette
}
