% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_silhouette.pdist.R
\name{silhouette.pdist}
\alias{silhouette.pdist}
\title{Pairwise \eqn{L_p} Distance of Multiple Persistence Silhouettes}
\usage{
silhouette.pdist(slist, p = 2, as.dist = FALSE)
}
\description{
Pairwise \eqn{L_p} Distance of Multiple Persistence Silhouettes
}
