% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.pdist2.R
\name{silhouette.pdist2}
\alias{silhouette.pdist2}
\title{Pairwise \eqn{L_p} Distances for Two Sets of Persistence Silhouettes}
\usage{
silhouette.pdist2(slist1, slist2, p = 2)
}
\description{
Pairwise \eqn{L_p} Distances for Two Sets of Persistence Silhouettes
}
