% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.wsum.R
\name{silhouette.wsum}
\alias{silhouette.wsum}
\title{Weighted Sum of Multiple Persistence Silhouettes}
\usage{
silhouette.wsum(slist, weight)
}
\description{
Weighted Sum of Multiple Persistence Silhouettes
}
