% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.eqdist.R
\name{test.eqdist}
\alias{test.eqdist}
\title{Energy-based Testing Equality of Distributions for Persistence Landscapes}
\usage{
test.eqdist(dlist, label, method = c("original", "discoB", "discoF"),
  mc.iter = 496)
}
\description{
Energy-based Testing Equality of Distributions for Persistence Landscapes
}
